compileError tests and macros
Error formatting and usage of `compileError` slightly changed.
In Dotty, there is a built-in `typeCheckErrors` method. This
method attempts to compile a literal String and returns all the
error messages encountered. Previously, the same functionality
required a macro. The Dotty compiler has this method hard-coded
in its internals, hence no macro is needed. So effectively,
`typeCheckErrors` is treated as any other method.

Also, `compileTimeOnly` test was removed. This is because
Dotty has a built-in `erased` keyword which achieves the same
thing and is tested as part of Dotty itself.