Implement `assertGoldenFile` and `assertGoldenLiteral` tests (#394)
This PR implements `assertGoldenLiteral` and `assertGoldenFile`, which
are assertions that not only compare the given values (literal v.s.
literal or string v.s. file contents respectively), but can take the
`UTEST_UPDATE_GOLDEN_TESTS` environment variable to automatically update
the source code and golden files to new values.

- `assertGoldenFile` is straightforward: if the contents of the
specified file does not match the given string,
`UTEST_UPDATE_GOLDEN_TESTS` makes utest over-write that file with the
given string.

- `assertGoldenLiteral` is more tricky: this takes a new
`utest.framework.GoldenFix.Span[T]` implicit conversion similar to
`sourcecode.Text`, but instead of just capturing the text contents of
the expression it captures the sourcefile and the start/end offset of
the literal within it. If the two values do not match,
`UTEST_UPDATE_GOLDEN_TESTS` makes utest use `utest.shaded.pprint.apply`
to pretty-print the given expression and the sourcefile/start/end
metadata to over-write the original source file in-place. Some extra
book-keeping is necessary to make sure the offsets line up even when
there are multiple edits in the same file, and to ensure the indentation
of start of the expression is preserved



Inspired by Jane Street's
https://blog.janestreet.com/the-joy-of-expect-tests/[What if writing
tests was a joyful experience?] blog post