Don't compare types with equals (#401)
Noticed because the annotation type does not arrive dealiased in
2.13.17.

Almost ran the test suite against latest Scala but I must juggle JVM
versions. I'll update if it doesn't verify the fix. (Edit: verified with
JVM 8, with Scala 2 published as 2.13.16. I noticed that if I apply the
edit and re-run
```
./mill 'utest.jvm[2.13.16].test'
```
I must git clean for it to pass.)

Fixes #397

Co-authored-by: Li Haoyi <haoyi.sg@gmail.com>